requirements.txt
# external requirements
click~=8.1.3
Sphinx
coverage
flake8
pandas~=2.0.1
matplotlib
seaborn
scipy
scikit-learn~=1.2.2
flask~=2.3.2
pyyaml~=6.0
numpy~=1.24.3
joblib~=1.2.0
dvc
python-dotenv
boto3